feat: A controller has been added to handle requests related to goal management. Methods for adding, deleting, editing goals, and viewing the list of goals have been implemented. Corresponding commands have been added to the main menu to access the goal management functionality. refactor: the User class has been refactored, resulting in the identifier type being changed from long to int.

+++ src/main/java/belousov/eu/exception/UserNotFoundException.java
@@ -7,7 +7,7 @@ public class UserNotFoundException extends RuntimeException {
         super(String.format("Пользователь с электронной почтой %s не найден", emial));
     }
 
-    public UserNotFoundException(long id) {
+    public UserNotFoundException(int id) {
         super(String.format("Пользователь с идентификатором %s не найден", id));
     }
 }
